+++ tests/e2e/matrix/internal/transport/admin/snapshot.go
@@ -6,7 +6,7 @@ import (
 	"github.com/craftgodotdev/craftgo/pkg/server"
 	"net/http"
 
-	service "github.com/craftgodotdev/craftgo/tests/e2e/matrix/internal/service/admin-service/ops/v2"
+	service "github.com/craftgodotdev/craftgo/tests/e2e/matrix/internal/service/admin"
 	"github.com/craftgodotdev/craftgo/tests/e2e/matrix/svccontext"
 )